backend/requirements.txt
django
djangorestframework
markdown
django-filter
python-dotenv
psycopg2-binary
torch
torchaudio
pyannote.audio
scikit-learn
numpy